om_dynamic_reports/requirements.txt
xlsxwriter>=3.1.0
python-dateutil>=2.8.2
